feat: quote exact input with v4

+++ src/interfaces/IMixedRouteQuoterV2.sol
@@ -29,7 +29,7 @@ interface IMixedRouteQuoterV2 {
 
     struct QuoteExactInputSingleV4Params {
         PoolKey poolKey;
-        uint128 exactAmount;
+        uint256 exactAmount;
         uint160 sqrtPriceLimitX96;
         bytes hookData;
     }
